Fix pppWDrawMatrixFrontLoop size and data types
- Change final storage to float instead of integer cast
- Eliminates complex float-to-int conversion (fctiwz/stfd/lwz/stw)
- Reduces function size from 148 to 136 bytes (matches target)
- Fixes stack frame size from 0x30 to 0x20 bytes
- Maintains proper PSMTXScaleApply and PSMTXMultVec calls

+++ src/pppWDrawMatrixFrontLoop.cpp
@@ -29,7 +29,7 @@ void pppWDrawMatrixFrontLoop(struct _pppPObject* param_1)
 	
 	PSMTXMultVec(ppvCameraMatrix0, &local_18, &local_18);
 	
-	*(s32*)((char*)param_1 + 0x4c) = (s32)local_18.x;
+	*(float*)((char*)param_1 + 0x4c) = local_18.x;
 	*(float*)((char*)param_1 + 0x5c) = local_18.y;
 	*(float*)((char*)param_1 + 0x6c) = local_18.z;
 }
